fix(dashboard): fix Next.js 15 catch-all route constraint violations
`[...name]/edit`, `[...name]/plan`, and `[...name]/run` all violated
Next.js 15's rule that catch-all segments must be the last URL segment.
The server logged "Catch-all must be the last part of the URL" and
refused to start fresh — the existing dev server was only serving
stale-compiled routes from before commit 1edf57bc renamed [name] to
[...name].

Fix strategy:
- Rename `edit/` → `_edit/` and `plan/` → `_plan/` (underscore prefix
  = private folder, not a route segment in the App Router).
- Convert `RecipeEditPage` and `RecipePlanPage` to accept `{ name }` prop
  instead of route `params`.
- Add a thin `RecipePage` router wrapper as the default export of
  `[...name]/page.tsx` that unwraps params, detects "edit"/"plan" as
  the last segment, and renders the correct sub-view — preserving all
  existing URLs.
- Fold the `/run` route into `[...name]/route.ts` with a last-segment
  check, replacing the deleted `run/route.ts` child route.

All 733 tests pass; dogfood confirmed 200 on /, /recipes/:name,
/recipes/:name/edit, /recipes/:name/plan, and the run API.
