+++ dashboard/src/app/api/bridge/recipes/[...name]/route.ts
@@ -13,6 +13,42 @@ export const runtime = "nodejs";
 // Next 15: dynamic route params arrive as a Promise.
 type RouteContext = { params: Promise<{ name: string[] }> };
 
+export async function POST(
+  req: NextRequest,
+  ctx: RouteContext,
+): Promise<Response> {
+  const guard = requireSameOrigin(req);
+  if (guard) return guard;
+  const { name } = await ctx.params;
+  if (name[name.length - 1] !== "run") {
+    return new Response(JSON.stringify({ error: "Method not allowed" }), {
+      status: 405,
+      headers: { "content-type": "application/json" },
+    });
+  }
+  const recipeName = name.slice(0, -1).join("/");
+  const read = await readBodyWithCap(req, BRIDGE_BODY_CAPS.run);
+  if (!read.ok) return bodyTooLargeResponse(BRIDGE_BODY_CAPS.run);
+  try {
+    const res = await bridgeFetch(`/recipes/${recipeName}/run`, {
+      method: "POST",
+      headers: { "content-type": req.headers.get("content-type") ?? "application/json" },
+      body: read.body || undefined,
+    });
+    const text = await res.text();
+    return new Response(text, {
+      status: res.status,
+      headers: { "content-type": "application/json" },
+    });
+  } catch (err) {
+    console.error("[recipes/:name/run] bridge fetch failed:", err);
+    return new Response(
+      JSON.stringify({ error: "Bridge unreachable" }),
+      { status: 502, headers: { "content-type": "application/json" } },
+    );
+  }
+}
+
 export async function GET(
   _req: NextRequest,
   ctx: RouteContext,
