+++ dashboard/src/app/recipes/[...name]/_edit/_components/YamlEditor.tsx
